// frontend/types/jobInfos.type.ts
export interface IJobInfo {
  department?: string;
  remainingEmployeeCount?: string;
  gender?: string;
  city?: string;
  location?: string;
  ageRange?: string;
  minDegree?: string;
  requiredDegree?: string;
  jobHistory?: string;
  clock?: string;
  salary?: string;
  mission?: string;
  description?: string;
}